Implement current hover tile job

+++ Assets/Scripts/Structs/SSoldier.cs
@@ -3,7 +3,7 @@ using Unity.Mathematics;
 
 public struct SSoldier : IComponentData
 {
-    public float2 currentCoordinates;
+    public int2 currentCoordinates;
     public int Movement;
     public int Initiative;
 }
